# swagger.yaml
swagger: '2.0'
info:
  version: '1.0'
  title: "is21"
  description: api for managing exams and students
  license:
    name: MIT
    url: https://github.com/apiaryio/polls-api/blob/master/LICENSE
host: polls.apiblueprint.org
basePath: /
schemes:
- http
consumes:
- application/json
produces:
- application/json
paths:
...
  /exams:
    get:
      summary: 'Get exams list'
      description: 'Get the exams list'
      responses:
        200:
          description: 'OK'
          schema:
            $ref: '#/definitions/ArrayOfExams'
        204:
          description: 'No exams'
    post:
      summary: 'Create exam'
      description: 'Create an exam with questions'
      parameters:
        - name: "properties"
          in: body
          required: true
          schema:
            $ref: '#/definitions/ExamData'
      responses:
        201:
          description: 'OK'
          schema:
            $ref: '#/definitions/Exam'
        400:
          description: "Error creating exams, wrong data"
        500:
          description: "Error creating exam in the database"
  /exams/{exam_id}:
    get:
      summary: 'Get exam'
      description: 'Get an exam tasks'
      parameters:
        - name: exam_id
          in: path
          required: true
          type: integer
      responses:
        200:
          description: 'OK'
          schema:
            $ref: '#/definitions/Exam'
        400:
          description: 'Exam not found, bad exam_id'
        404:
          description: 'Exam not found'
    patch:
      summary: 'Update exam'
      description: 'Update exam tasks by id'
      parameters:
        - name: exam_id
          in: path
          required: true
          type: integer
        - name: exam_update
          in: body
          required: true
          schema:
            $ref: '#/definitions/ExamData'
      responses:
        200:
          description: 'Exam updated'
          schema:
            $ref: '#/definitions/Exam'
        400:
          description: 'Exam not found, bad exam_id'
        404:
          description: 'Exam not found'

    delete:
      summary: 'Delete an exam'
      description: 'Delete an existing exam by id'
      parameters:
        - name: exam_id
          in: path
          required: true
          type: integer
      responses:
        204:
          description: 'Exam deleted'
        400:
          description: 'Exam not found, bad exam_id'
        404:
          description: 'Exam not found'
  /exams/{exam_id}/marks:
    get:
      summary: 'Get exam marks'
      description: 'Get all marks from an exam'
      parameters:
        - name: 'exam_id'
          in: path
          required: true
          type: integer
      responses:
        200:
          description: 'OK'
          schema:
            $ref: '#/definitions/ArrayOfMarks'
        204:
          description: 'No submissions for the exam'
        400:
          description: 'Bad exam_id'
        401:
          description: 'Unauthorized'
        404:
          description: 'Exam not found'
  /exams/{owner_id}/owner:
    get:
      summary: 'List exams created by ownerId'
      description: 'List exams created by ownerId'
      parameters:
        - name: owner_id
          in: path
          required: true
          type: integer
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/ArrayOfExams'
        204:
          description: 'No exams'
        400:
          description: 'Bad owner_id'
        404:
          description: 'Owner not found'
        401:
          description: 'Unauthorized'
...
definitions:
...
  Task:
    type: object
    properties:
      id:
        type: integer
      owner_id:
        type: integer
      task_type:
        enum: [SINGLE, MULTIPLE, OPEN]
      text:
        type: string
        example: 'Quanto fa 2+2?'
      aswers_id:
        type: integer
...
  Exam:
    type: object
    properties:
      id:
        type: integer
      owner_id:
        type: integer
      questions:
        type: array
        items:
          $ref: '#/definitions/Task'
  ExamData: #exam without the id (needed when creating the exam)
    type: object
    properties:
      owner_id:
        type: integer
      questions:
        type: array
        items:
          $ref: '#/definitions/Task'
...
  Mark:
    type: integer
  ArrayOfMarks:
    type: array
    items:
        $ref: '#/definitions/Mark'
  ArrayOfExams:
    type: array
    items:
        $ref: '#/definitions/Exam'
...